% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{consistency.of.phenotypes}
\alias{consistency.of.phenotypes}
\title{Consistency of phenotypes plot}
\usage{
consistency.of.phenotypes(object, entityName, geneName, type = "box", ...)
}
\arguments{
\item{object}{An RNAi object.}

\item{entityName}{A character string giving the name of the entity in question, e.g. "BONE".}

\item{geneName}{A character string giving the name of the gene in question, e.g. "MDM4".}

\item{type}{The style of plot one wishes to show, can be one of "box" or "line".}

\item{...}{Any other additions to the ggplot. One can change the theme, plot name, etc. as desired.}
}
\value{
A ggplot object.
}
\description{
Generate a plot comparing the LFC values of each shRNA of a given gene for each cell line of a given entity.
}
\examples{
data("Achilles2.4_Data")
data("Achilles2.4_annotation")

RNAiObject <- new(Class = "RNAi", genes = Achilles2.4_Data$Gene,
                  sequences = Achilles2.4_Data$Name,
                  cancers = colnames(Achilles2.4_Data)[-(1:2)],
                  cancersP53 = Achilles2.4_annotation,
                  values = Achilles2.4_Data[,-(1:2)])

consistency.of.phenotypes(object = RNAiObject, entityName = "BONE", geneName = "MDM4", type = "line")

}

