% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{p53.dependency.by.entity}
\alias{p53.dependency.by.entity}
\title{P53 dependency by entity}
\usage{
p53.dependency.by.entity(object, entityName, geneName, p53 = "all",
  type = "all", ...)
}
\arguments{
\item{object}{An RNAi object.}

\item{entityName}{A character string giving the name of the entity in question, e.g. "BONE".}

\item{geneName}{A character string giving the name of the gene in question, e.g. "MDM4".}

\item{p53}{The types of cell lines to include. Can be "all", "MUT", "WT", or "NA".}

\item{type}{The style of plot one wishes to show, can be one of "box" or "line".}

\item{...}{Any other additions to the ggplot. One can change the theme, plot name, etc. as desired.}
}
\value{
A ggplot object.
}
\description{
Generate a plot comparing the dependency of various shRNAs on the mutation status of p53 for the cells lines of a given entity.
}
\examples{
data("Achilles2.4_Data")
data("Achilles2.4_annotation")

RNAiObject <- new(Class = "RNAi", genes = Achilles2.4_Data$Gene,
                  sequences = Achilles2.4_Data$Name,
                  cancers = colnames(Achilles2.4_Data)[-(1:2)],
                  cancersP53 = Achilles2.4_annotation,
                  values = Achilles2.4_Data[,-(1:2)])

p53.dependency.by.entity(object = RNAiObject, entityName = "Lung", geneName = "MDM4", p53 = "all", type = "line")

}

