% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{riger.gene.comparison.datasets}
\alias{riger.gene.comparison.datasets}
\title{RIGER gene score comparison for two datasets}
\usage{
riger.gene.comparison.datasets(object1, object2, object1Name = "Dataset1",
  object2Name = "Dataset2", geneName = "MDM4", entityName = "all",
  percentage = 5, plotName = "Dataset Comparisson Plot", ...)
}
\arguments{
\item{object1}{A RIGER type object, see vignette for details.}

\item{object2}{A RIGER type object, see vignette for details.}

\item{object1Name}{A character string giving the name of object1.}

\item{object2Name}{A character string giving the name of object2.}

\item{geneName}{A character string giving the name of the gene in question, e.g. "MDM4".}

\item{entityName}{A character string giving the name of the entity in question, e.g. "BONE". Can also be "all".}

\item{percentage}{The percentage of scores which should be considered toxic or enriching.}

\item{plotName}{A character string gicing the name of the final plot.}

\item{...}{Any other additions to the ggplot. One can change the theme, plot name, etc. as desired.}
}
\value{
A ggplot object.
}
\description{
Generate a plot comparing the LFC values of each shRNA of a given gene within entities for two different datasets.
}
\examples{
data("Achilles2.0")
data("Achilles2.4")

riger.gene.comparison.datasets(object1 = Achilles2.0, object2 = Achilles2.4,
                               object1Name = "Ach1", object2Name = "Ach2",
                               geneName = "MDM4",entityName = "Colon",
                               percentage = 1, plotName = "Comparing RIGER scores")

}

