% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{riger.gene.comparison}
\alias{riger.gene.comparison}
\title{RIGER gene score comparison}
\usage{
riger.gene.comparison(object, geneName, entityName = "all", ...)
}
\arguments{
\item{object}{A RIGER type object, see vignette for details.}

\item{geneName}{A character string giving the name of the gene in question, e.g. "MDM4".}

\item{entityName}{A character string giving the name of the entity in question, e.g. "BONE". Can also be "all".}

\item{...}{Any other additions to the ggplot. One can change the theme, plot name, etc. as desired.}
}
\value{
A ggplot object.
}
\description{
Generate a plot comparing the LFC values of each shRNA of a given gene within entities. The entity scores are calculated based on the mean or median of the relevant cell line scores.
}
\examples{
data("Achilles2.4")

riger.gene.comparison(object = Achilles2.4, geneName = "MDM4", entityName = "all")

}

