% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{variation.within.cell.line}
\alias{variation.within.cell.line}
\title{Variation within cell line}
\usage{
variation.within.cell.line(object, cancerID, geneName, type = "all", ...)
}
\arguments{
\item{object}{An RNAi object.}

\item{cancerID}{a character string indicating the cell line of interest (partial string matching is sufficient).}

\item{geneName}{A character string giving the name of the gene in question, e.g. "MDM4".}

\item{type}{The style of plot one wishes to show, can be one of "all" or "median".}

\item{...}{Any other additions to the ggplot. One can change the theme, plot name, etc. as desired.}
}
\value{
A ggplot object.
}
\description{
Generate a plot comparing the LFC values of each shRNA of a given gene within a chosen cell line.
}
\examples{
data("Achilles2.4_Data")
data("Achilles2.4_annotation")

RNAiObject <- new(Class = "RNAi", genes = Achilles2.4_Data$Gene,
                  sequences = Achilles2.4_Data$Name,
                  cancers = colnames(Achilles2.4_Data)[-(1:2)],
                  cancersP53 = Achilles2.4_annotation,
                  values = Achilles2.4_Data[,-(1:2)])

variation.within.cell.line(object = RNAiObject, cancerID = "SLR24", geneName = "MDM4", type = "all")

}

