% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{variation.within.entities}
\alias{variation.within.entities}
\title{Variation within entities}
\usage{
variation.within.entities(object, geneName, type = "value",
  statistic = "mean", ...)
}
\arguments{
\item{object}{An RNAi object.}

\item{geneName}{A character string giving the name of the gene in question, e.g. "MDM4".}

\item{type}{The type of values one wishes to show. Either "rank" or "value" (for the LFC scores).}

\item{statistic}{A statistic used to group the cell line scores, can either be "mean" or "median".}

\item{...}{Any other additions to the ggplot. One can change the theme, plot name, etc. as desired.}
}
\value{
A ggplot object.
}
\description{
Generate a plot comparing the LFC values of each shRNA of a given gene within all entities. The entity scores are calculated based on the mean or median of the relevant cell line scores.
}
\examples{
data("Achilles2.4_Data")
data("Achilles2.4_annotation")

RNAiObject <- new(Class = "RNAi", genes = Achilles2.4_Data$Gene,
                  sequences = Achilles2.4_Data$Name,
                  cancers = colnames(Achilles2.4_Data)[-(1:2)],
                  cancersP53 = Achilles2.4_annotation,
                  values = Achilles2.4_Data[,-(1:2)])

variation.within.entities(object = RNAiObject, geneName = "MDM4", type = "value", statistic = "mean")

}

